refactor(M6 PR2): extract column/tree mouse handling into controller/mouse.rs
Move handle_mouse + the non-modal column/tree pointer handling (wheel scroll,
divider/scrollbar drag with click-to-scroll, click vs double-click, the
track->offset rounding, hit_test) into a submodule, along with the
is_double_click timing rule + DOUBLE_CLICK and the timing unit test. The
shared scroll/nav primitives and the MouseRegion/Drag types stay in mod.rs
(Drag backs a struct field). finder imports the shared is_double_click.
Behavior-preserving. mod.rs ~1940.
